requirements.txt
openai
GoogleBard
termcolor
requests